[Hermes/AST] populate declaration lists of external ASTs
Summary:
When using our own parser, we adds the declarations to the appropriate
AST declaration list. However those lists are empty if we are using an
external AST (like the one provided by the Flow parser).

So, in this commit populate the declaration lists of the AST.

Test Plan: ninja check-hermes

Reviewers: #hermes

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D17649381