[Hermes] options to dump lazily generated code
Summary:
If any of the `--dump-xxx` options is used together with `--lazy`, we
use that as the dumping behavior of lazy compilation, instead of eager
compilation. So, we can finally see what lazy compilation is generating.

I was afraid that calling the bytecode disassembler from the VM might
increase the binary size, so `--lazy --dump-bytecode` only works in
debug builds.

Test Plan:
  hermes --Xlazy1 --lazy --dump-ir test.js
  hermes --Xlazy1 --lazy --dump-bytecode test.js
and so on.

Reviewers: #hermes

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D19307382