[Hermes] add hidden options for lazy thresholds
Summary:
Sometimes it is convenient to tweak the lazy compilation thresholds for
testing. Add three hidden options:

`--Xlazy-function-threshold`
In lazy mode a function source must be at least this many bytes long in
order to be parsed lazily. Smaller functions are always parsed eagerly.

`--Xlazy-unit-threshold`
In lazy mode the input to the compiler must be at least this many bytes
long in order to be parsed lazily.

`--Xlazy1`
A shortcut setting both options above to 1.

Note that these options a hidden and unsupported, as indicated by the
traditional prefix `-X`. They are only intended for testing and thus are
global. They are not per instance of Hermes!

Test Plan: Experimented with options and compilation of small functions.

Reviewers: #hermes

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D19286956