[Hermes] Object scoping compiler support
Summary:
Add the optional object scoping mode to the compiler. When enabled,
variables are stored in ordinary objects linked by their prototypes
instead of environments. This simplifies lazy compilation, debugging,
local eval.

For fully defined scope chains, most variable instructions are simply
lowered to existing object and property access instructions.
`CreateScope` creates a new object, `LoadVariable` reads a property, and
so on. Other new instructions emit new bytecodes:

- `GetParentScope` emits `GetObjectParent` in object scoping mode.
- `ReadOnlyVariable` emits `ReadOnlyProp` in object scoping mode and
  nothing otherwise.

Additionally, the case where the scope chain is not fully defined must
be supported. When accessing a variable not defined in the enclosing
lexical scopes, the compiler emits `LoadDynamic`/`StoreDynamic`, which
are lowered to `GetDynamic`/`PutDynamic`. These instruction lookup
variables by name. The variable may be declared in unknown enclosing
scopes, in the global scope, or it may be missing.

Object scoping is enabled with `-fobject-scoping`, which enables support
for the "with" statement.

As an optimization use the closest scope when accessing a variable.
Previously we would always start from the current scope when accessing a
variable, regardless of the scope of the variable itself. In many cases
the variable's scope may already be available in a register, or at least
a scope closer to it may be available.

Use the closest to the desired scope instead of the current one.
Test Plan: Sandcastle

Reviewers: #hermes

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D25853805