[HermesVM] Define StaticScope and DynamicScope cells
Summary:
StaticScope and DynamicScope represent the new object scopes at
runtime. StaticScope is a lexical scope with a fixed set of variables,
while DynamicScope is either the scope of a non-strict local eval(), or
the scope of a "with" statement.

Both are thin subclasses of JSObject. DynamicScope adds a member with an
optional pointer to the "with" object.

The parent scopes are linked through the prototype chain, but in the
future, when we move the prototype in the hidden class, they will have
to be linked through an explicit pointer. Otherwise we would need a new
hidden class for every inner function invocation.

Test Plan: at the end of the stack

Reviewers: #hermes

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D25941926