[HermesVM] add a new "global cache"
Summary:
The global cache is similar to the existing property cache, but is a
little (very little) more general purpose, shared, fixed size, and
aliased.

Reliably keying instructions in the cache is tricky, because an
instruction address in memory does not uniquely identify an instruction.
First, one instruction can belong to multiple functionsi, which execute
in different lexical context (because of bytecode de-duplication), and
second, bytecode can be added to the VM at runtime, as well as freed.

So, to uniquely identify a bytecode instruction, we use three pieces of
data:
- a 48-bit unique RuntimeModule id (a per-runtime counter)
- a 16-bit function ID
- the instruction IP

It is possible to implement property caching on top of it, but it would
be less efficient because of aliasing and contention for the same cache
entry between different instructions.

The goal of this is to enable caching in new kinds of instructions
without having to redesign CodeBlock.

The cache is mapped instead of malloc-ed, so it won't consume memory
until the first time it is actually used.

The immediate goal for the cache is to be used by the dynamic scoping
instructions introduced in the next commit.

Test Plan: In the next commits

Reviewers: #hermes

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D25853803