[HermesVM] New property flag throwOnWrite
Summary:
When using object scoping, at runtime we need to be able to
differentiate between ES6 `const` and an ES5-style read-only variable.
ES6 `const` always throws on modification, while ES5-style throws only
in strict mode.

We are introducing a new specialized property flag `throwOnWrite`. It
can only be used in local object scopes, which implies that the
properties are "ordinary". It can only be set when `writable` is cleared
and it indicates that any write to that property must throw, even in
non-strict mode.

Regular JSObject functions do not support this flag - they cannot set it
and do not check it. It can only be set by a new JSObject method
`updateOwnPropertyToConst()`. The flag is only checked by the
instructions `PutDynamic` and `TryPutDynamic` in a later diff.

Test Plan: later in the stack

Reviewers: avp, Hermes

Subscribers:

Tasks:

Tags: