[Hermes] explicit scopes
Summary:
Start managing lexical scopes explicitly in the generated code.
Previously scopes were implicit, and there was only a single scope per
function. With the new model, scopes are created explicitly with a new
instruction and there can be multiple ones in a single function.

This finally enables us to generate correct ES6 code, but also fixes
code generation for ES5 `catch(e)`, which also requires a new scope.
Additionally, it enables a new and much simpler approach to lazy
compilation.

Because of the scope (pun intended) of the change, this diff doesn't
address everything. Debugging is disabled, many optimizations are
disabled, lazy compilation is disabled. Even with optimizations, there
are expected to be some bytecode regressions because some bytecode
instructions are becoming larger. All of these will be addressed in
subsequent commits which will all be landed together.

Some old IR values and instructions have not been deleted yet to
decrease the changes.

Two new IR values are being introduced:

### ScopeDesc
This is the "descriptor" of a scope, containing a list of `ScopeVar`
instances and a pointer to a parent scope. This is "meta data" - it has
no representation at runtime.

It is passed as an additional parameter to IR instructions returning a
scope, so the compiler can keep track of which scope it is.

### ScopeVar
This is the replacement of `Variable`. It is a variable that belongs to
a `ScopeDesc`.

The following new IR instructions are added:

### GetFunctionParentScope scopeDesc
Returns the scope in which the current function was defined.

### CreateScope parentScope, scopeDesc
Create a new scope. `parentScope` is the value holding the parent scope
or `undefined`, and `scopeDesc` is the descriptor of the new scope.

### GetParentScope startScope, startDesc, desiredDesc
Starting from the specified scope, walk the parents until the desired
scope is reached. This allows us to access variables in parent scopes.
`startDesc` and `desiredDesc` encode the metadata of how many scopes up
to go.

### LoadVariable var, startScope, startDesc
Load the specified variable. `startScope` and `startDesc` encode the
current scope. The variable may belong to a parent scope.

### StoreVariable value, targetVal, startScope, startDesc
The counterpart to `LoadVariable`.

### DeclareGlobalVariable
Declare a new global variable. Previously this was done implicitly
during instruction selection by scanning the function metadata.

Test Plan: test262

Reviewers: #hermes

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D25817468